requirements.txt
langchain
langchain-openai
chromadb
pydantic
streamlit
pandas
python-dotenv
tiktoken
pymupdf
langchain-community
sqlalchemy
